Organize platform logic and consolidate common alias handling
