+++ lib/commands/alias.js
@@ -1,4 +1,4 @@
-const PlatformStorage = require('../platform-storage');
+const PlatformStorage = require('../platform/platform-storage');
 
 function listAliases(aliases) {
   const moduleNames = Object.getOwnPropertyNames(aliases);
@@ -17,31 +17,23 @@ function aliasCommand(argv) {
 
   switch (argv.sub) {
     case 'add':
-      platform.addAlias(argv.mod, argv.path)
-        .then(() => {
-          console.log(`Alias for ${argv.mod} added to platform.`);
-        })
-        .catch((error) => {
-          console.log('Unable to add alias:', error);
-        });
+      platform.addAlias(argv.mod, argv.path);
+      console.log(`Alias for ${argv.mod} added to platform.`);
       break;
     case 'remove':
-      platform.removeAlias(argv.mod)
-        .then(() => {
-          console.log(`Alias for ${argv.mod} added to platform.`);
-        })
-        .catch((error) => {
-          console.log('Unable to remove alias:', error);
-        });
+      if (platform.hasAlias(argv.mod)) {
+        platform.removeAlias(argv.mod);
+        console.log(`Alias for ${argv.mod} removed from platform.`);
+      } else {
+        console.log(`Platform does not contain module ${argv.mod}`);
+      }
       break;
     case 'clear':
-      platform.clearAliases()
-        .then(() => {
-          console.log('All aliases have been removed from platform.');
-        });
+      platform.clearAliases();
+      console.log('All aliases have been removed from platform.');
       break;
     case 'list':
-      platform.getAllAliases().then(listAliases);
+      listAliases(platform.getAllAliases());
       break;
     default:
       console.log('No sub-command invoked.');
