+++ lib/commands/build.js
@@ -1,6 +1,6 @@
 const context = require('../cli-context');
 const stripes = require('@folio/stripes-core/webpack/stripes-node-api');
-const StripesPlatform = require('../stripes-platform');
+const StripesPlatform = require('../platform/stripes-platform');
 const { applyOptions, stripesConfigOptions, buildOptions } = require('./common-options');
 const { processError, processStats, cliResolve, cliAliases } = require('../webpack-common');
