+++ lib/commands/serve.js
@@ -1,6 +1,6 @@
 const context = require('../cli-context');
 const stripes = require('@folio/stripes-core/webpack/stripes-node-api');
-const StripesPlatform = require('../stripes-platform');
+const StripesPlatform = require('../platform/stripes-platform');
 const { applyOptions, serverOptions, stripesConfigOptions, buildOptions } = require('./common-options');
 const { cliResolve, cliAliases } = require('../webpack-common');
