+++ lib/commands/status.js
@@ -1,7 +1,7 @@
 const context = require('../cli-context');
 const packageJson = require('../../package.json');
-const StripesPlatform = require('../stripes-platform');
-const PlatformStorage = require('../platform-storage');
+const StripesPlatform = require('../platform/stripes-platform');
+const PlatformStorage = require('../platform/platform-storage');
 const { applyOptions, stripesConfigOptions } = require('./common-options');
 
 function statusCommand(argv) {
