+++ lib/commands/test.js
@@ -1,7 +1,7 @@
 const context = require('../cli-context');
 const stripes = require('@folio/stripes-core/webpack/stripes-node-api');
 const integrationTests = require('../integration-tests');
-const StripesPlatform = require('../stripes-platform');
+const StripesPlatform = require('../platform/stripes-platform');
 const { applyOptions, serverOptions, stripesConfigOptions } = require('./common-options');
 const { cliAliases, cliResolve } = require('../webpack-common');
