+++ lib/stripes-cli.js
@@ -4,6 +4,7 @@ const yargs = require('yargs');
 const updateNotifier = require('update-notifier');
 const packageJson = require('../package.json');
 const context = require('./cli-context');
+const { AliasError } = require('./platform/alias-validation');
 
 process.title = 'stripes-cli';
 
@@ -17,8 +18,16 @@ updateNotifier({
 
 console.log(`\nstripes-cli ${packageJson.version}\n`);
 
-const argv = yargs.commandDir('./commands')
-  .example('$0 <command> --help', 'View examples and options for a command')
-  .demandCommand()
-  .help()
-  .argv;
+try {
+  const argv = yargs.commandDir('./commands')
+    .example('$0 <command> --help', 'View examples and options for a command')
+    .demandCommand()
+    .help()
+    .argv;
+} catch (err) {
+  if (err instanceof AliasError) {
+    console.log(`Alias Error: ${err.message}`);
+  } else {
+    throw err;
+  }
+}
